tickable blocks
world now tracks a list of tickable blocks in world->ticks

+++ src/block_types/plant.c
@@ -1,10 +1,11 @@
 #include <stdlib.h>
+#include <stdio.h>
 #include "plant.h"
 #include "../utils.h"
 
 void plant_tick(plant_t *plant, double time) {
 	if (!d_close(plant->growth, plant->fullgrown)) {
-		float next_growth = plant->growth + time * plant->growth_rate;	
+		double next_growth = plant->growth + time * plant->growth_rate;	
 		plant->growth = MIN(next_growth, plant->fullgrown);
 	}
 }
